Namespace the blob reaper as Images::CleanUnattachedBlobsJob
