+++ app/jobs/bike_jobs/remove_orphaned_images_job.rb
@@ -19,7 +19,7 @@ module BikeJobs
         1.week.ago..1.day.ago
       end
 
-      # The stamp is what exempts a blob from CleanUnattachedBlobsJob, so it has to be what
+      # The stamp is what exempts a blob from Images::CleanUnattachedBlobsJob, so it has to be what
       # collects it here too - anything it skips that this misses leaks forever
       def blobs_for(stolen_record_id)
         ActiveStorage::Blob.where("binx_data->>'stolen_record_id' = ?", stolen_record_id.to_s)
