+++ app/jobs/scheduled_job_runner.rb
@@ -41,13 +41,13 @@ class ScheduledJobRunner < ScheduledJob
       BikeJobs::UpdateTheftAlertFacebookJob,
       CleanBParamsJob,
       CleanBulkImportJob,
-      CleanUnattachedBlobsJob,
       CreateGraduatedNotificationJob,
       CreateStolenGeojsonJob,
       CreateUserAlertNotificationJob,
       Email::ScheduledSurveyJob,
       FetchProject529BikesJob,
       FileCacheMaintenanceJob,
+      Images::CleanUnattachedBlobsJob,
       ImpoundExpirationJob,
       ProcessGraduatedNotificationJob,
       ProcessHotSheetJob,
