+++ spec/jobs/images/clean_unattached_blobs_job_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe CleanUnattachedBlobsJob, type: :job do
+RSpec.describe Images::CleanUnattachedBlobsJob, type: :job do
   include_context :scheduled_job
   include_examples :scheduled_job_tests
